Implement the scripting API in imwarp itself.
