+++ dipy/align/imwarp.py
@@ -12,6 +12,8 @@ from dipy.align import floating
 from dipy.align import VerbosityLevels
 from dipy.align import Bunch
 from dipy.data import get_data
+from dipy.align.metrics import CCMetric, EMMetric, SSDMetric
+
 
 RegistrationStages = Bunch(INIT_START=0,
                            INIT_END=1,
@@ -1722,49 +1724,50 @@ class SymmetricDiffeomorphicRegistration(DiffeomorphicRegistration):
         return self.static_to_ref
 
 
-def get_synthetic_warped_circle(nslices):
+metric_dict = {'CC': CCMetric, 
+               'EM': EMMetric, 
+               'SSD': SSDMetric}
+
+
+def syn_registration(moving, static, moving_grid2world=None, static_grid2world=None, 
+                     metric='CC', dim=3, level_iters = [10, 10, 5], prealign=None):
     """
-    Create synthetic data to test registration algorithms
+    Register a source image (moving) to a target image (static)
 
     Parameters
     ----------
-    nslices : int
-        The number of slices in the volume
+    moving : ndarray
+        The source image data to be registered
+    moving_grid2world : array, shape (4,4)
+        The affine matrix associated with the moving (source) data.
+    static : ndarray
+        The target image data for registration
+    static_grid2world : array, shape (4,4)
+        The affine matrix associated with the static (target) data
+    metric : string, optional
+        The metric to be optimized. One of `CC`, `EM`, `SSD`, Default: CCMetric.
+    dim: int (either 2 or 3), optional
+       The dimensions of the image domain. Default: 3
+    level_iters : list of int, optional
+        the number of iterations at each level of the Gaussian Pyramid (the
+        length of the list defines the number of pyramid levels to be
+        used).
 
     Returns
     -------
-    circle_3d : 3D array, shape (64, 64, nslices)
-        A volume with a cylinder
+    warped_moving : ndarray 
+        The data in `moving`, warped towards the `static` data.
 
-    wcircle_3d : 3D array, shape (64, 64, nslices)
-        A volume with the same cylinder warped through a DiffeomorphicMap
+    forward : ndarray (..., 3)
+        The vector field describing the forward warping from the source to the target. 
+    backward : ndarray (..., 3)
+        The vector field describing the backward warping from the target to the source
     """
-    #get a subsampled circle
-    fname_cicle = get_data('reg_o')
-    circle = np.load(fname_cicle)[::4,::4].astype(floating)
+    use_metric = metric_dict[metric](dim)
 
-    #create a synthetic invertible map and warp the circle
-    d, dinv = vfu.create_harmonic_fields_2d(64, 64, 0.1, 4)
-    d = np.asarray(d, dtype=floating)
-    dinv = np.asarray(dinv, dtype=floating)
-    mapping = DiffeomorphicMap(2, (64, 64))
-    mapping.forward, mapping.backward = d, dinv
-    wcircle = mapping.transform(circle)
+    sdr = SymmetricDiffeomorphicRegistration(use_metric, level_iters)
+    mapping = sdr.optimize(static, moving, static_grid2world=static_grid2world, 
+                            moving_grid2world=moving_grid2world, prealign=prealign)
 
-    if(nslices == 1):
-        return circle, wcircle
-
-    #normalize and form the 3d by piling slices
-    circle = (circle-circle.min())/(circle.max() - circle.min())
-    circle_3d = np.ndarray(circle.shape + (nslices,), dtype=floating)
-    circle_3d[...] = circle[...,None]
-    circle_3d[...,0] = 0
-    circle_3d[...,-1] = 0
-
-    #do the same with the warped circle
-    wcircle = (wcircle-wcircle.min())/(wcircle.max() - wcircle.min())
-    wcircle_3d = np.ndarray(wcircle.shape + (nslices,), dtype=floating)
-    wcircle_3d[...] = wcircle[...,None]
-    wcircle_3d[...,0] = 0
-    wcircle_3d[...,-1] = 0
-    return circle_3d, wcircle_3d
+    warped_moving = mapping.transform(moving)
+    return warped_moving, np.array(mapping.forward), np.array(mapping.backward)
